+++ src/JavaScriptRuntime/Modules/Shared/LocalModulesAssembly.cs
@@ -1,8 +1,0 @@
-using System.Reflection;
-
-namespace JavaScriptRuntime.Modules.Shared;
-
-class LocalModulesAssembly
-{
-    public Assembly? ModulesAssembly { get; set; }
-}